remove UT: home-page banner has a homepage link
Wanted to check that a hyperlink like `<a href="/" ...>Home</a>` was
present in the banner of the home page. It wasn't possible to include
this test in a way that wasn't frail: the values in "..." are important
when using `self.assertContains(response, something, html=True)`. This
means that any class/style modifications to the link would break the
test. This is better tested at functional-test level.